requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
pillow>=9.0.0
flask>=2.0.0
flask-cors>=3.0.10
